+++ src/lib/parse.ts
@@ -1,7 +1,9 @@
-import { getMonthLength } from './date-utils.js'
+import { getMonthLength } from '$lib/date-utils.js'
+import { type Locale, type InnerLocale, getInnerLocale } from '$lib/locale'
 
 type RuleToken = {
 	id: string
+	allowedValues?: string[]
 	toString: (d: Date) => string
 }
 
@@ -11,6 +13,7 @@ type ParseResult = {
 	date: Date | null
 	missingPunctuation: string
 }
+
 /** Parse a string according to the supplied format tokens. Returns a date if successful, and the missing punctuation if there is any that should be after the string */
 export function parse(str: string, tokens: FormatToken[], baseDate: Date | null): ParseResult {
 	let missingPunctuation = ''
@@ -54,6 +57,20 @@ export function parse(str: string, tokens: FormatToken[], baseDate: Date | null)
 		}
 	}
 
+	function parseEnum(allowedValues: string[]) {
+		const n = allowedValues.findIndex((allowedValue) => {
+			return allowedValue.toLowerCase() === str.slice(0, allowedValue.length).toLowerCase()
+		})
+
+		if (n >= 0) {
+			str = str.slice(allowedValues[n].length)
+			return n
+		} else {
+			valid = false
+			return null
+		}
+	}
+
 	function parseToken(token: FormatToken) {
 		if (typeof token === 'string') {
 			parseString(token)
@@ -66,6 +83,9 @@ export function parse(str: string, tokens: FormatToken[], baseDate: Date | null)
 		} else if (token.id === 'MM') {
 			const value = parseUint(/^[0-9]{2}/, 1, 12)
 			if (value !== null) month = value - 1
+		} else if (token.id === 'MMM') {
+			const value = parseEnum(token.allowedValues || [])
+			if (value !== null) month = value
 		} else if (token.id === 'dd') {
 			const value = parseUint(/^[0-9]{2}/, 1, 31)
 			if (value !== null) day = value
@@ -101,48 +121,56 @@ function twoDigit(value: number) {
 	return ('0' + value.toString()).slice(-2)
 }
 
-const ruleTokens: RuleToken[] = [
-	{
-		id: 'yyyy',
-		toString: (d: Date) => d.getFullYear().toString(),
-	},
-	{
-		id: 'yy',
-		toString: (d: Date) => d.getFullYear().toString().slice(-2),
-	},
-	{
-		id: 'MM',
-		toString: (d: Date) => twoDigit(d.getMonth() + 1),
-	},
-	{
-		id: 'dd',
-		toString: (d: Date) => twoDigit(d.getDate()),
-	},
-	{
-		id: 'HH',
-		toString: (d: Date) => twoDigit(d.getHours()),
-	},
-	{
-		id: 'mm',
-		toString: (d: Date) => twoDigit(d.getMinutes()),
-	},
-	{
-		id: 'ss',
-		toString: (d: Date) => twoDigit(d.getSeconds()),
-	},
-]
-function parseRule(s: string) {
-	for (const token of ruleTokens) {
-		if (s.startsWith(token.id)) {
-			return token
+function parseRule(s: string, innerLocale: InnerLocale) {
+	if (s.startsWith('yyyy')) {
+		return {
+			id: 'yyyy',
+			toString: (d: Date) => d.getFullYear().toString(),
+		}
+	} else if (s.startsWith('yy')) {
+		return {
+			id: 'yy',
+			toString: (d: Date) => d.getFullYear().toString().slice(-2),
+		}
+	} else if (s.startsWith('MMM')) {
+		return {
+			id: 'MMM',
+			allowedValues: innerLocale.shortMonths,
+			toString: (d: Date) => innerLocale.shortMonths[d.getMonth()],
+		}
+	} else if (s.startsWith('MM')) {
+		return {
+			id: 'MM',
+			toString: (d: Date) => twoDigit(d.getMonth() + 1),
+		}
+	} else if (s.startsWith('dd')) {
+		return {
+			id: 'dd',
+			toString: (d: Date) => twoDigit(d.getDate()),
+		}
+	} else if (s.startsWith('HH')) {
+		return {
+			id: 'HH',
+			toString: (d: Date) => twoDigit(d.getHours()),
+		}
+	} else if (s.startsWith('mm')) {
+		return {
+			id: 'mm',
+			toString: (d: Date) => twoDigit(d.getMinutes()),
+		}
+	} else if (s.startsWith('ss')) {
+		return {
+			id: 'ss',
+			toString: (d: Date) => twoDigit(d.getSeconds()),
 		}
 	}
 }
 
-export function createFormat(s: string): FormatToken[] {
+export function createFormat(s: string, locale: Locale = {}): FormatToken[] {
+	const innerLocale = getInnerLocale(locale)
 	const tokens = []
 	while (s.length > 0) {
-		const token = parseRule(s)
+		const token = parseRule(s, innerLocale)
 		if (token) {
 			// parsed a token like "yyyy"
 			tokens.push(token)
@@ -157,5 +185,6 @@ export function createFormat(s: string): FormatToken[] {
 			s = s.slice(1)
 		}
 	}
+
 	return tokens
 }