+++ src/routes/DatePicker.svelte
@@ -6,25 +6,41 @@
 
 	// had to import it this way to avoid errors
 	// in `npm run build:site` or `npm run check`:
-	import hy from 'date-fns/locale/hy/index.js'
+	import hy from 'date-fns/locale/hy/index'
+	import de from 'date-fns/locale/de/index'
+	import nb from 'date-fns/locale/nb/index'
+
 	let value: Date
 	let min: Date
 	let max: Date
-	let locale = localeFromDateFnsLocale(hy)
+	let locales = [
+		{ key: 'default', value: localeFromDateFnsLocale({}) },
+		{ key: 'nb (date-fns)', value: localeFromDateFnsLocale(nb) },
+		{ key: 'de (date-fns)', value: localeFromDateFnsLocale(de) },
+		{ key: 'hy (date-fns)', value: localeFromDateFnsLocale(hy) },
+	]
+	let locale = locales[0]
 	let browseWithoutSelecting: boolean
 	let timePrecision: 'minute' | 'second' | 'millisecond' | null = 'millisecond'
 </script>
 
 <Split>
 	<div class="left" slot="left">
-		<DatePicker bind:value bind:min bind:max {locale} bind:browseWithoutSelecting {timePrecision} />
+		<DatePicker
+			bind:value
+			bind:min
+			bind:max
+			locale={locale.value}
+			bind:browseWithoutSelecting
+			{timePrecision}
+		/>
 	</div>
 	<div slot="right">
 		<h3 class="no-top">Props</h3>
 		<Prop label="value">{value}</Prop>
 		<Prop label="min" bind:value={min} />
 		<Prop label="max" bind:value={max} />
-		<Prop label="locale">date-fns <code>hy</code></Prop>
+		<Prop label="locale" bind:value={locale} values={locales} />
 		<Prop label="browseWithoutSelecting" bind:value={browseWithoutSelecting} />
 		<Prop
 			label="timePrecision"